Mass Update
-Changelog 24/11/19
-Question 5

+++ Ecommerce/Ecommerce/order.cpp
@@ -1,12 +1,13 @@
 #include "order.h"
 //----------------------------------------------------------------------------------------//
-Order::Order(Products** product_list, const int list_size, const int total_price)
+Order::Order(Products** product_list, const int list_size, const float total_price)
 {
 	m_product_list = product_list;
 	setListSize(list_size);
 	setTotalPrice(total_price);
 }
 //----------------------------------------------------------------------------------------//
+/*
 Order::Order(const Order& other) //copy c'tor
 {
 	m_product_list = other.m_product_list;
@@ -19,6 +20,7 @@ Order::Order(Order&& other) //move c'tor
 	other.m_product_list = nullptr;
 	m_total_price = other.m_total_price;
 }
+*/
 //----------------------------------------------------------------------------------------//
 Order::~Order()
 {
@@ -27,7 +29,7 @@ Order::~Order()
 	delete[] m_product_list;
 }
 //----------------------------------------------------------------------------------------//
-bool Order::setTotalPrice(const int total_price)
+bool Order::setTotalPrice(const float total_price)
 {
 	m_total_price = total_price;
 }
@@ -42,7 +44,7 @@ const Products** Order::getProductList() const
 	return m_product_list;
 }
 //----------------------------------------------------------------------------------------//
-const int Order::getTotalPrice() const
+const float Order::getTotalPrice() const
 {
 	return m_total_price;
 }
